PhysioRepo started. PhysioAuthenticateLogin created.

+++ GTFys/Models/TreatmentType.cs
@@ -12,10 +12,10 @@ namespace GTFys.Models
         // Private properties representing attributes of a treatment type
         private string Type { get; set; }       // The type or name of the treatment
         private TimeSpan Time { get; set; }     //The duration of the treatment (represented as a TimeSpan)
-        private decimal Cost { get; set; }      // The cost of the treatment
+        private double Cost { get; set; }      // The cost of the treatment
 
         // Constructor to initialize a TreatmentType object
-        public TreatmentType(string type, TimeSpan time, decimal cost)
+        public TreatmentType(string type, TimeSpan time, double cost)
         {
             // Set the private properties of the TreatmentType object based on the provided parameters
             Type = type;
